requirements-ci.txt
ansible-lint==25.8.1
black==25.1.0
flake8
flake8-docstrings
flake8-import-order
pytest==8.4.1
reorder-python-imports
